Added BFQ and SIO I/O Scheduler
Add the BFQ-v6r2 I/O scheduler to 3.1.
The general structure is borrowed from CFQ, as much code. A (bfq_)queue
is
associated to each task doing I/O on a device, and each time a
scheduling
decision has to be made a queue is selected and served until it expires.